\name{SPAP-package}
\alias{SPAP-package}
\alias{SPAP}
\docType{package}
\title{
  Shortest Path Algorithms
}
\description{
  This package has the Dijksta and Bellman-Ford algorithms, it also constructs a custom class "graph" that stores the user's graph easily.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Robert Horne, roberthorne@outlook.com
}
\keyword{ Shortest Path Algotithms, Dijksta, Bellman-Ford }
